### 2021-05-12 #3275 @whedon
[PRE REVIEW]: PyEI: A Python package for ecological inference
**Submitting author:** @karink520 (<a href="http://orcid.org/0000-0003-1312-6473">Karin Knudson</a>)
**Repository:** <a href="https://github.com/mggg/ecological-inference" target ="_blank">https://github.com/mggg/ecological-inference</a>
**Version:** v0.1.0
