+++ paper/paper.bib
@@ -1,4 +1,5 @@
 @article{goodman1953ecological,
+  doi={10.2307/2088121},
   title={Ecological regressions and behavior of individuals.},
   author={Goodman, Leo A},
   journal={American sociological review},
@@ -106,6 +107,7 @@ Estimates},
   }
   
   @book{elliot2000spatial,
+  doi={10.1093/acprof:oso/9780198515326.001.0001},
   title={Spatial epidemiology: methods and applications.},
   author={Elliot, Paul and Wakefield, Jon C and Best, Nicola G and Briggs, David John and others},
   year={2000},
@@ -139,6 +141,7 @@ Survey Information'},
 }
 
 @article{neal2011mcmc,
+  doi={10.1201/b10905-7}
   title={MCMC using Hamiltonian dynamics},
   author={Neal, Radford},
   journal={Handbook of markov chain monte carlo},
