+++ paper/paper.bib
@@ -141,7 +141,7 @@ Survey Information'},
 }
 
 @article{neal2011mcmc,
-  doi={10.1201/b10905-7}
+  doi={10.1201/b10905-7},
   title={MCMC using Hamiltonian dynamics},
   author={Neal, Radford},
   journal={Handbook of markov chain monte carlo},
