### 2021-05-13 #3275 @karink520
Thanks, @danielskatz! I updated the DOIs. There is one that is still listed as missing. It looks to me like the suggestion is a DOI for data that accompanies the book, but that there is no DOI for the book itself, so I will leave that as is if that works.
