### 2021-06-07 7fbc2fa
Add first-draft of margin plot function
