### 2021-06-09 a349b94
fix threshold verbose reporting
