### 2021-06-22 #3275 @drvinceknight
@eigenfoo @ejhigson @pmyteh @SuperKam91 and  @matt-graham I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "PyEI: A Python package for ecological inference". You can find more information at the top of this Github issue.
