### 2021-06-23 #3397 @whedon
[REVIEW]: PyEI: A Python package for ecological inference
**Submitting author:** @karink520 (<a href="http://orcid.org/0000-0003-1312-6473">Karin Knudson</a>)
**Repository:** <a href="https://github.com/mggg/ecological-inference" target ="_blank">https://github.com/mggg/ecological-inference</a>
**Version:** v0.1.1
