### 2021-07-08 mggg/ecological-inference#60 @matt-graham
HTML API documentation
Submitting as part of JOSS review openjournals/joss-reviews#3397

Although there is API documentation in the docstrings in the code, there is no associated autogenerated HTML documentation which makes it a little difficult to browse. A nice addition would be to have autogenerated API documentation hosted on for example GitHub pages or Read the Docs.
