### 2021-07-08 mggg/ecological-inference#61 @matt-graham
Expanding introduction at beginning of README
Submitting as part of JOSS review openjournals/joss-reviews#3397

While the introduction notebooks do give a good summary of what problems the software is designed to solve and who the target audience is, the introduction at the beginning of the README file is currently a bit lacking in detail. I would suggest adding a couple of sentences defining what ecological inference is  and what functionality PyEI offers.
