### 2021-07-08 mggg/ecological-inference#62 @matt-graham
Missing argument descriptions in some docstrings
Submitting as part of JOSS review openjournals/joss-reviews#3397

Some functions and methods do not have any descriptions of the arguments in their docstrings, for example

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L27-L33
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L219-L222
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L277-L279
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L326-L333
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L382
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/two_by_two.py#L548-L551
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/goodmans_er.py#L21
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/goodmans_er.py#L116
https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/pyei/r_by_c.py#L135

If any of the (non-`__init__`) functions / methods are intended for internal use only it may be worth indicating that by prefixing the names with an underscore and/or explicitly declaring the public API of each module in the `__all__` attribute.
