### 2021-07-08 mggg/ecological-inference#63 @matt-graham
Convergence issues in `PyEI_overview` notebook
Submitting as part of JOSS review openjournals/joss-reviews#3397

A potential concern I have about the package is that by hiding away to some extent the details of how inference is performed, there is a danger of a user analyzing the results without being aware of the implications of potential convergence issues in the Markov chain Monte Carlo sampling. This is not a purely hypothetical scenario, as the chains sampled when fitting the example model in the [`pyei/intro_notebooks/PyEI_overview.ipynb`](https://github.com/mggg/ecological-inference/blob/main/pyei/intro_notebooks/PyEI_overview.ipynb) notebook appear to be suffering convergence issues.

In particular the chains appear to include a non-negligible number of [divergent transitions](https://mc-stan.org/docs/2_19/reference-manual/divergent-transitions) in the post-warmup stage, which is generally indicative of a problematic posterior geometry with varying scales / curvature. The R̂ convergence diagnostics estimates computed by ArviZ  &mdash; 1.09 for `phi_1` and 1.07 for `phi_2` on a run with `random_seed=20210708` &mdash; also indicate that the chains are poorly converged, as do the low effective sample size estimates.

While warning messages from PyMC3 which highlight these issues are displayed, there is no discussion in the notebook of these issues and possible remedies. Further as the `target_accept` parameter is already set to the very high value of 0.99, there is limited scope for further increases as recommended in the warning messages. I think it would be worthwhile to include at least some brief discussion of these issues in the notebook(s) with pointers to relevant resources in the PyMC3 or Stan documentation. Ideally there would also be an example of experimentation with approaches to mitigate these issues, for example, using additional prior knowledge to define more informative priors which might shift the posterior mass away from the problematic regions; trying alternative parametrizations - the docstring for `ei_beta_binom_model_modified` suggests this is already a reparametrization of the standard formulation - if the standard formulation does indeed suffer even more from convergence issues it could be worthwhile including this as an initial example in the notebook to illustrate the importance of the choice of parametrization.
