### 2021-07-08 mggg/ecological-inference#64 @matt-graham
`lint_and_test.sh` script not running tests
Submitting as part of JOSS review openjournals/joss-reviews#3397

The script `scripts/lint_and_test.sh` currently does not run the tests on my machine (running code as of  2d47262e413a4e17db1eead0924cff164e57f01b on Ubuntu 20.04). It seems that the `pylint` check may be returning an error status which due to the `set -ex` call then causes the script to exit without running the `scripts/test.sh` script. Running `scripts/test.sh` directly works however.
