### 2021-07-08 mggg/ecological-inference#65 @matt-graham
No mention of possibility of installing directly from PyPI
Submitting as part of JOSS review openjournals/joss-reviews#3397

While the package has releases on `PyPi` the installation instructions suggest to install directly from GitHub rather than running `pip install pyei`. Is this intentional? If you want the default to be installing from the latest version on the `main` branch it may still be worth adding a pointer to the PyPI version as an alternative less up-to-date but potentially more stable version.
