### 2021-07-08 mggg/ecological-inference#66 @matt-graham
Enhancing `conda` environment set up instructions
Submitting as part of JOSS review openjournals/joss-reviews#3397

In the 'Install with conda' instructions in the README file the `conda install pip` line is (I think) unnecessary as I think installing `python` in a `conda` environment will also always install `pip`. All of the requirements also seem to available as `conda` packages if the `conda-forge` channel is used, so to [avoid mixing `conda` and `pip`](https://www.anaconda.com/blog/using-pip-in-a-conda-environment) an alternative to create the environment and install all of the dependencies would be
```
conda create --name pyei --channel conda-forge python=3.8 --file requirements.txt --file requirements-dev.txt
```
