### 2021-07-08 mggg/ecological-inference#67 @matt-graham
Table in paper included as image rather than text
Submitting as part of JOSS review openjournals/joss-reviews#3397

In the following

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L37

the table is included as an image. Ideally the table should be text based rather than an image to make it more accessible to readers using assistive technology and also searchable. I am not sure if it is possible to create a table with a caption in the Markdown syntax used to generate JOSS papers &ndash; the JOSS documentation suggests Pandoc is used to generate the PDF and the [Pandoc documentation](https://pandoc.org/MANUAL.html#tables) does suggest there is support for tables with captions though this requires an extension and I am not sure if this is enabled or not.
