### 2021-07-08 mggg/ecological-inference#68 @matt-graham
Use of i.i.d. in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L68

'i.i.d.' should be defined on first usage, also independent may be better as 'i.i.d. samples from the posterior distribution' is somewhat redundant as 'from the posterior distribution' already implies the samples are identically distributed.
