### 2021-07-08 mggg/ecological-inference#69 @matt-graham
Description of effective sample size in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L68

In the description of effective sample size (lines 62--65 in proof PDF), I think it would be better to refer to estimates of posterior expectations in general rather than the posterior mean in particular, as effective sample sizes can and commonly are computed for other quantities such as the posterior variances. For example, something like: 'Consider effective sample size, which is a measure of how the variance of a Monte Carlo estimate of a posterior expectation computed from dependent samples compares to the variance of a corresponding estimate computed from independent samples (or, very roughly, how "effective" the samples are for estimating a posterior expectation compared to independent samples)'.
