### 2021-07-08 mggg/ecological-inference#70 @matt-graham
Clarification of assumptions required for dimension scaling results in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L68

The claim regarding the scaling with dimension for Metropolis&ndash;Hastings versus Hamiltonian Monte Carlo (in the sentence in lines 65&ndash;69 in proof PDF) should ideally have a clarification that the result requires assumptions on the target distribution (in particular that it is a product of independent factors in each dimension) to hold. To avoid having to introduce technical details this could just simply be adding 'Under certain assumptions on the target posterior distribution,' at the beginning of the sentence.
