### 2021-07-08 mggg/ecological-inference#71 @matt-graham
Incorrect capitlization in references in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

Some capitalization in the references in `paper/paper.bib` needs fixing by adding braces `{}`, for example, 'bayesian' &rightarrow; 'Bayesian', 'python' &rightarrow; 'Python', 'goodman' &rightarrow; 'Goodman', 'hamiltonian' &rightarrow; 'Hamiltonian, 'monte' &rightarrow; 'Monte', 'carlo' &rightarrow; 'Carlo'.
