### 2021-07-08 mggg/ecological-inference#72 @matt-graham
Reducing number of example plots in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

While it is nice to have a couple of example outputs, I feel that five is possibly excessive. The example notebooks already provide a venue for giving a more exhaustive tour of the plotting functionality while I would say the aim in the paper is just to support the assertions in paragraph in lines 75--76 in the proof PDF. To me selecting one or two key plots and including in the text rather than at the end would be preferable.
