### 2021-07-08 mggg/ecological-inference#73 @matt-graham
KDE not defined in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L83

The acronym KDE in the caption should be defined.
