### 2021-07-08 mggg/ecological-inference#74 @matt-graham
Adding further reference for MCMC algorithm used in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L68

The default sampler in PyMC3, as with Stan, I believe uses an updated algorithm that includes several improvements over the original NUTS algorithm. [Betancourt (2017)](https://arxiv.org/abs/1701.02434) describes the main improvements and so would be a worthwhile additional reference in line 60 in the proof PDF.
