### 2021-07-08 mggg/ecological-inference#75 @matt-graham
Removing parentheses around paragraph in paper
Submitting as part of JOSS review openjournals/joss-reviews#3397

https://github.com/mggg/ecological-inference/blob/2d47262e413a4e17db1eead0924cff164e57f01b/paper/paper.md#L61

The parentheses around this paragraph (lines 45&ndash;47 in the proof PDF) is, I think, unnecessary and interrupts the flow of the paper.
