### 2021-07-08 #3397 @matt-graham
I have added the bulk of my review as issues on the project repository; most of these should only require small fixes. The main substantive issues I would say are some possible improvements to the API documentation (mggg/ecological-inference#60 and mggg/ecological-inference#62) and adding some discussion of MCMC convergence issues to the notebooks (mggg/ecological-inference#63).

I think overall this submission has a lot of strong points. It seems like a nicely designed Python package which provides a useful contribution to the Python ecosystem by increasing the ease of performing (Bayesian) ecological inference. The provided example notebooks are a great resource for understanding how to use the package to perform inference and create visualisations of the results. The code has decent test coverage, with continuous integration via Github Actions. The paper provides a good summary of the need for the software and its functionality, and gives pointers to comparable packages in R. As far as I can tell from some brief searching there are no other comparable Python packages which have been omitted, or similar packages from other languages such as MATLAB or Julia. The stated functionality all seems to be present in the software.

One general question I have is about the list of authors in the paper. The JOSS reviewer guidelines [request that reviewers check that the list of authors appears complete](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) and raise a question if this is not obviously the case. From looking through the commit history it seems that @ColCarroll has made non-negligible contributions to the codebase. Although they are included in the acknowledgements it feels they could also validly be included as an author. Can I check therefore if @ColCarroll has been consulted about authorship?
