### 2021-07-08 #3397 @ColCarroll
Hey @matt-graham -- the authors talked to me beforehand. My contributions were around testing/CI (which got me on the commit history!), as well as advising on some computation and API issues, and I am delighted to be in the acknowledgements. Thank you for checking on it, too!
