### 2021-07-14 mggg/ecological-inference#76 @pmyteh
Nested parens in citation
Submitting as part of JOSS review openjournals/joss-reviews#3397

In the paper, one of the references comes out as `(see especially (King, 1997))` with nested parentheses. As the JOSS format is (I think) RMarkdown, you can avoid this by putting the parenthetical text in the citation square brackets: `[see especially @king1997solution]` or whatever.
