### 2021-07-14 mggg/ecological-inference#77 @pmyteh
Import issue: update requirements.txt
Submitting as part of JOSS review openjournals/joss-reviews#3397

I start this issue by _expressly_ declaring that it is an upstream problem (and not therefore the fault of the authors). It also doesn't occur in a clean venv. Nevertheless, there is a versioning issue with `pymc3` that can make `import pyei` fail in some circumstances:

```
Python 3.9.2 (default, Feb 28 2021, 17:03:44)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.20.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import pyei
<ipython-input-1-2a13c949293a> in <module>
----> 1 import pyei

~/.local/lib/python3.9/site-packages/pyei/__init__.py in <module>
      1 """A package for rpv and ecological inference"""
      2 __version__ = "0.1.0"
----> 3 from .two_by_two import *
      4 from .goodmans_er import *
      5 from .plot_utils import *

~/.local/lib/python3.9/site-packages/pyei/two_by_two.py in <module>
      8
      9 import warnings
---> 10 import pymc3 as pm
     11 import numpy as np
     12 import theano.tensor as tt

~/.local/lib/python3.9/site-packages/pymc3/__init__.py in <module>
     39
     40 from .blocking import *
---> 41 from .distributions import *
     42 from .distributions import transforms
     43 from .glm import *

~/.local/lib/python3.9/site-packages/pymc3/distributions/__init__.py in <module>
     13 #   limitations under the License.
     14
---> 15 from . import timeseries
     16 from . import transforms
     17 from . import shape_utils

~/.local/lib/python3.9/site-packages/pymc3/distributions/timeseries.py in <module>
     21
     22 from pymc3.util import get_variable_name
---> 23 from .continuous import get_tau_sigma, Normal, Flat
     24 from .shape_utils import to_tuple
     25 from . import multivariate

~/.local/lib/python3.9/site-packages/pymc3/distributions/continuous.py in <module>
     30 from .special import log_i0
     31 from ..math import invlogit, logit, logdiffexp
---> 32 from .dist_math import (
     33     alltrue_elemwise, betaln, bound, gammaln, i0e, incomplete_beta, logpow,
     34     normal_lccdf, normal_lcdf, SplineWrapper, std_cdf, zvalue,

~/.local/lib/python3.9/site-packages/pymc3/distributions/dist_math.py in <module>
     26 from theano.scalar import UnaryScalarOp, upgrade_to_float_no_complex
     27 from theano.tensor.slinalg import Cholesky
---> 28 from theano.scan_module import until
     29 from theano import scan
     30 from .shape_utils import to_tuple

ModuleNotFoundError: No module named 'theano.scan_module'
```
This project's `requirements.txt` pins version 3.9.3 of `pymc3`. That version [has a requirement](https://github.com/pymc-devs/pymc3/blob/v3.9.3/requirements.txt) of `theano>=1.05`. Although 1.0.5 was the last version of the original `theano`, development has continued as [`aesara`](https://github.com/aesara-devs/aesara), which continues to install as `theano` but has releases up to 2.0.5. And version 1.0.11 [reorganised the `scan_module` module](https://github.com/aesara-devs/aesara/releases/tag/rel-1.0.11) such that `pymc3` 3.9.3 no longer loads.

The issue is fixed in later version of `pymc3`, so if you can loosen this package's pin on `pymc3` then this issue will go away. If not, perhaps include `theano>=1.05,<1.11` in your own `requirements.txt`.
