### 2021-07-14 mggg/ecological-inference#77 @pmyteh
(Although on reflection I don't know if there's anything else in `theano` versions after 1.0.5 that might be problematic. That version is known-working, so perhaps a simple `theano==1.05` pin would be more reliable).
