### 2021-07-14 mggg/ecological-inference#61 @pmyteh
Agreed. The brief description at the top of the paper is very good and could usefully be repurposed.
