### 2021-07-15 mggg/ecological-inference#63 @pmyteh
I agree. I was pleased to see a brief discussion at the bottom of `Introduction_to_EI.ipynb` about things that can go wrong with Bayesian modelling, and that PyEI was 'more vocal' about revealing such issues to the user. Brilliant! That these issues arise and are reported by the package when running the examples provides a wonderful opportunity to highlight the relative seriousness of these issues _in the context of EI work_. You (the authors) will know *far* more about what is concerning and what is not than the majority of your userbase will.

Obviously there's something of a balance: package documentation can't replace a hefty stats textbook. But maybe a short overview of the warnings issued and pointers to interpreting them in the overview docs, together with a brief statement of why you consider the model 'good enough' to proceed despite the warnings in the example notebooks. And given that, frankly, many users will be trying these models _without_ reading the textbook first, heading them off from the obvious footguns would be a serious form of academic service!
