### 2021-07-15 mggg/ecological-inference#60 @pmyteh
I would put this as a 'nice-to-have' rather than a blocker for acceptance. But it *would* be nice to have!
