### 2021-07-15 #3397 @pmyteh
This is an excellent submission for JOSS. The package meets a real need (ecological inference within the Python ecosystem) and is well executed. The design seems sound, the tutorial and example notebooks are very welcome, and the code is clean and consistent. The paper gives a good overview and is engagingly written. Kudos to the authors.

@matt-graham has already identified most of the issues I spotted. I have +1'd a number of key issues raised by him, and added comments to a couple of them. I've also opened two new issues: one trivial and one more important (though probably rare in practice and very easy to fix). My only other nice-to-have would be to add comments to suppress pylint for the two remaining `not-context-manager` false-positives found by `lint-and-test.py` (as well as fixing it to actually trigger the tests as noted in issue 64).

Again - well done. And thanks for giving me a chance to review it.
