### 2021-07-15 mggg/ecological-inference#60 @matt-graham
Agreed - this is in reality more a nice-to-have than a blocker for acceptance and I should have made that clearer by not referencing it directly in the review issue - I will update the issue to reflect this.
