+++ paper/paper.bib
@@ -8,7 +8,7 @@
 }
 
 @article{greiner2009r,
-  title={R$\times$ C ecological inference: bounds, correlations, flexibility and transparency of assumptions},
+  title={{R}$\times$ {C} ecological inference: bounds, correlations, flexibility and transparency of assumptions},
   author={James Greiner, D and Quinn, Kevin M},
   journal={Journal of the Royal Statistical Society: Series A (Statistics in Society)},
   volume={172},
@@ -48,7 +48,7 @@
 }
 
 @article{rosen2001bayesian,
-  title={Bayesian and frequentist inference for ecological inference: The R$\times$ C case},
+  title={Bayesian and frequentist inference for ecological inference: The {R}$\times$ {C} case},
   author={Rosen, Ori and Jiang, Wenxin and King, Gary and Tanner, Martin A},
   journal={Statistica Neerlandica},
   volume={55},
@@ -67,12 +67,12 @@
     number = {33},
     pages = {1143},
     author = {Ravin Kumar and Colin Carroll and Ari Hartikainen and Osvaldo Martin},
-    title = {ArviZ a unified library for exploratory analysis of Bayesian models in Python},
+    title = {{ArviZ} a unified library for exploratory analysis of {Bayesian} models in {Python}},
     journal = {Journal of Open Source Software}
 }
 
 @article{salvatier2016probabilistic,
-  title={Probabilistic programming in Python using PyMC3},
+  title={Probabilistic programming in {Python} using PyMC3},
   author={Salvatier, John and Wiecki, Thomas V and Fonnesbeck, Christopher},
   journal={PeerJ Computer Science},
   volume={2},
@@ -82,7 +82,7 @@
 }
 
 @Manual{eiCompare,
-    title = {eiCompare: Compares Ecological Inference, Goodman, Rows by Columns
+    title = {{eiCompare}: Compares Ecological Inference, {Goodman}, Rows by Columns
 Estimates},
     author = {Loren Collingwood and Ari Decter-Frain and Hikari Murayama and Pratik Sachdeva and Juandalyn Burke},
     year = {2020},
@@ -91,7 +91,7 @@ Estimates},
   }
 
   @Manual{ei,
-    title = {ei: Ecological Inference},
+    title = {{ei}: Ecological Inference},
     author = {Gary King and Molly Roberts},
     year = {2016},
     note = {R package version 1.3-3},
@@ -99,7 +99,7 @@ Estimates},
   }
   
     @Manual{eiPack,
-    title = {eiPack: Ecological Inference and Higher-Dimension Data Management},
+    title = {{eiPack}: Ecological Inference and Higher-Dimension Data Management},
     author = {Olivia Lau and Ryan T. Moore and Michael Kellermann},
     year = {2020},
     note = {R package version 0.2-1},
@@ -115,7 +115,7 @@ Estimates},
 }
     
     @Manual{RxCEcolInf,
-    title = {RxCEcolInf: 'R x C Ecological Inference With Optional Incorporation of
+    title = {RxCEcolInf: '{R} x {C} Ecological Inference With Optional Incorporation of
 Survey Information'},
     author = {D. James Greiner and Paul Baines and Kevin M. Quinn},
     year = {2019},
@@ -124,7 +124,7 @@ Survey Information'},
   }
     
     @article{hoffman2014no,
-  title={The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo.},
+  title={The {No-U-Turn} sampler: adaptively setting path lengths in {Hamiltonian} {Monte} {Carlo}.},
   author={Hoffman, Matthew D and Gelman, Andrew},
[…]
