### 2021-07-15 9d60d56
changing iid to independent, since it is clear that it refers to samples are all from the (identical) posterior dist
