### 2021-07-15 261c36a
removing parentheses around paragraph
