### 2021-07-15 712cd97
adding a more detailed introduction in the README
