+++ paper/paper.bib
@@ -151,6 +151,14 @@ Survey Information'},
   year={2011}
 }
 
+@misc{betancourt2018conceptual,
+      title={A Conceptual Introduction to Hamiltonian Monte Carlo}, 
+      author={Michael Betancourt},
+      year={2018},
+      eprint={1701.02434},
+      archivePrefix={arXiv},
+      primaryClass={stat.ME}
+}
 
     
     
