### 2021-07-19 mggg/ecological-inference#64 @karink520
Ah, thanks! Looks like pylint shifted under us, so we hadn't been seeing this error before. Will update.
