### 2021-07-19 320efd7
disabling not-context-manager check from pylint
