### 2021-07-19 43faff4
revise conda installation instructions to avoid mixing conda and pip
