### 2021-07-20 9520ad9
assumptions for dimension scaling results
