+++ pyei/two_by_two.py
@@ -30,6 +30,20 @@ def truncated_normal_asym(
     """
     A modification of king97's truncated normal that puts some broad priors
     over the parameters of the truncated normal dist
+
+    Parameters
+    ----------
+    group_fraction: Length-p (p=# of precincts) vector giving demographic information
+        as the fraction of precinct_pop in the demographic group of interest
+    votes_fraction: Length p vector giving the fraction of each precinct_pop that
+        votes for the candidate of interest
+    precinct_pops: Length-p vector giving size of each precinct population of interest
+         (e.g. voting population)
+
+    Returns
+    -------
+    model: A pymc3 model
+
     """
     num_precincts = len(precinct_pops)
     b_1_l_bound = np.maximum(0, (votes_fraction - 1 + group_fraction) / group_fraction)
@@ -219,6 +233,26 @@ def ei_beta_binom_model(group_fraction, votes_fraction, precinct_pops, lmbda):
 def log_binom_sum(lower, upper, obs_vote, n0_curr, n1_curr, b_1_curr, b_2_curr, prev):
     """
     Helper function for computing log prob of convolution of binomial
+
+    Parameters
+    ----------
+    lower, upper : lower and upper bounds on the (unobserved) count of votes from given
+        deographic group for given candidate within precinct
+    n0_curr: the (current value for the) count of given demographic group in the precinct
+    n1_curr: the (current value for the)count of the complement of given demographic group
+         in the precinct
+    b_1_curr: corresponds to p0 in wakefield's notation, the probability that an individual
+        in the given demographic group votes for the given candidate 
+    b_2_curr: corresponds to p1 in wakefield's notation, the probability that an individual
+        in the complement of the given demographic group votes for the given candidate
+    prev: the partial sum of the log prob of the convolution of binomials
+
+    Returns
+    -------
+    prev + component_for_current_precinct : sum of the previous value of the partial sum
+        for the log prob of the convolution of binomials and an additional term for one
+        precinct
+
     """
 
     # votes_within_group_count is y_0i in Wakefield's notation, the count of votes from
@@ -278,8 +312,20 @@ def wakefield_model_beta(
     group_fraction, votes_fraction, precinct_pops, pareto_scale=8, pareto_shape=2
 ):
     """
-    2 x 2 EI model Wakefield
+    2 x 2 EI model based on Wakefield's, with pareto distributions in upper level of hierarchy
 
+    Parameters
+    ----------
+    group_fraction: Length-p (p=# of precincts) vector giving demographic information
+        as the fraction of precinct_pop in the demographic group of interest
+    votes_fraction: Length p vector giving the fraction of each precinct_pop that
+        votes for the candidate of interest
+    precinct_pops: Length-p vector giving size of each precinct population of interest
+         (e.g. voting population)
+
+    Returns
+    -------
+    model: A pymc3 model
     """
 
     vote_count_obs = votes_fraction * precinct_pops
@@ -329,6 +375,24 @@ def wakefield_normal(group_fraction, votes_fraction, precinct_pops, mu0=0, mu1=0
 
     Note: Wakefield suggests adding another level of hierarchy, with a prior over mu0 and mu1,
     sigma0, sigma1, but that is not yet implemented here
+
+    Parameters
+    ----------
+    group_fraction: Length-p (p=# of precincts) vector giving demographic information
+        as the fraction of precinct_pop in the demographic group of interest
+    votes_fraction: Length p vector giving the fraction of each precinct_pop that
+        votes for the candidate of interest
+    precinct_pops: Length-p vector giving size of each precinct population of interest
+         (e.g. voting population)
+    mu0: float
+        Mean of the normally distributed hyperparameter associated with the demographic
[…]
