### 2021-07-25 fc11ed2
include table as markdown instead of image
