### 2021-07-27 c9acc2e
updating docstrings
