### 2021-08-03 mggg/ecological-inference#77 @karink520
Thanks for this -- I hadn't come across it yet! I've pinned to theano==1.0.5 for now; we're hoping  to loosen the pymc3 pin in the future.
