### 2021-08-03 ad51b5b
adding info about convergence issues to overview notebook
