### 2021-08-05 mggg/ecological-inference#63 @matt-graham
The changes made in ad51b5b look great to me and resolve this from my perspective, so closing.
