### 2021-08-05 mggg/ecological-inference#69 @matt-graham
Changes to text made in 9430426 look good to me so closing.
