### 2021-08-05 mggg/ecological-inference#67 @matt-graham
Fixed by fc11ed207d3bb0963c6731dafa0c203b8c3398d5 so closing.
