### 2021-08-05 mggg/ecological-inference#62 @matt-graham
Docstrings updated in c9acc2e0b01284d34dfe0971b71b2473ad78145c so closing.
