### 2021-08-10 #3397 @drvinceknight
@karink520 could you take a look at the missing DOIs? Whedon has suggested one of them but there are others that are missing:

- [ ] Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis. CRC press.
- [ ] Hoffman, M. D., & Gelman, A. (2014). The No-U-Turn sampler: Adaptively setting path lengths in Hamiltonian Monte Carlo. J. Mach. Learn. Res., 15(1), 1593–1623.
- [ ] James Greiner, D., & Quinn, K. M. (2009). R× C ecological inference: Bounds, correlations, flexibility and transparency of assumptions. Journal of the Royal Statistical Society: Series A (Statistics in Society), 172(1), 67–81.
- [ ] King, G. (1997). A solution to the ecological inference problem: Reconstructing individual behavior from aggregate data. Princeton University Press.
- [ ] King, G., Rosen, O., & Tanner, M. A. (1999). Binomial-beta hierarchical models for ecological inference. Sociological Methods & Research, 28(1), 61–90.
- [ ] Rosen, O., Jiang, W., King, G., & Tanner, M. A. (2001). Bayesian and frequentist inference for ecological inference: The R× C case. Statistica Neerlandica, 55(2), 134–156
- [ ] Salvatier, J., Wiecki, T. V., & Fonnesbeck, C. (2016). Probabilistic programming in Python using PyMC3. PeerJ Computer Science, 2, e55.
- [ ] Wakefield, J. (2004). Ecological inference for 2× 2 tables. Journal of the Royal Statistical Society: Series A (Statistics in Society), 167(3), 385–425.
