### 2021-08-11 #3397 @karink520
@drvinceknight Thanks! All the references we could find DOIs for are now updated. (The one that Whedon suggested has not been updated, since the suggestion is the DOI for an associated dataset, but I don't believe there is a DOI for the book itself).
