### 2021-08-13 #3397 @drvinceknight
> since the suggestion is the DOI for an associated dataset, but I don't believe there is a DOI for the book itself).

Could you double check if this is the DOI for the book itself (it's unclear to me if it's for the book or the chapter or just completely incorrect all together): https://doi.org/10.1515/9781400849208.343
