+++ pyei/__init__.py
@@ -1,5 +1,5 @@
 """A package for rpv and ecological inference"""
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 from .two_by_two import *
 from .goodmans_er import *
 from .plot_utils import *
